+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+# [1.0.4](https://github.com/stauren/vite-plugin-deadfile/tree/v1.0.4) (2023-11-22)
+- using @rollup/pluginutils to handle "include" and "exclude" config as file filter
+- add a new config `includeHiddenFiles`
+- fix a bug when using relative path as `root` (by @arnriu)
+- merge the plugin's first param `projectRoot` in to the config object. (by @arnriu)
+
 # [1.0.3](https://github.com/stauren/vite-plugin-deadfile/tree/v1.0.3) (2023-11-16)
 
 - fix a missing dist file bug